notes/requirements.txt
Django==3.1.4
gunicorn==20.0.4
psycopg2-binary==2.9.1
djangorestframework==3.12.1